Move LegalityChecker code to LegalityPanel
